1) app version number variable changed to "long long" 2) improved title handling for action-sheet 3) improved cleaning for older app versions

+++ infrared_ios/Library/Components/App/IRAppDescriptor.h
@@ -13,7 +13,7 @@
 @interface IRAppDescriptor : IRBaseDescriptor
 
 @property (nonatomic, readonly) NSString *app;
-@property (nonatomic) NSInteger version;
+@property (nonatomic) long long version;
 @property (nonatomic) BOOL silentUpdate;
 @property (nonatomic, strong) NSString *baseUrl;
 @property (nonatomic, strong) NSArray *screensArray;
